+++ src/Core/Domain/Configuration/User/Model/User.php
@@ -23,10 +23,7 @@ declare(strict_types=1);
 
 namespace Core\Domain\Configuration\User\Model;
 
-use Centreon\Domain\Common\Assertion\Assertion;
-use Core\Contact\Domain\Model\ContactTemplate;
-
-class User
+class User extends NewUser
 {
     public const MIN_ALIAS_LENGTH = 1,
                  MAX_ALIAS_LENGTH = 255,
@@ -40,203 +37,30 @@ class User
                  THEME_DARK = 'dark';
 
     /**
-     * @var integer|null
-     */
-    private ?int $id = null;
-
-    /**
-     * @var bool
-     */
-    private bool $isActivate = true;
-
-    /**
-     * @var bool
-     */
-    private bool $isAdmin = false;
-
-    /**
-     * @var string
-     */
-    private string $theme = self::THEME_LIGHT;
-
-    /**
-     * @var ContactTemplate|null
-     */
-    private ?ContactTemplate $contactTemplate = null;
-
-    /**
+     * @param int $id
      * @param string $alias
      * @param string $name
      * @param string $email
+     * @param bool $isAdmin
+     * @param string $theme
      * @throws \Assert\AssertionFailedException
      */
     public function __construct(
-        private string $alias,
-        private string $name,
-        private string $email,
+        protected int $id,
+        protected string $alias,
+        protected string $name,
+        protected string $email,
+        protected bool $isAdmin,
+        protected string $theme,
     ) {
         $this->setAlias($alias);
         $this->setName($name);
         $this->setEmail($email);
+        $this->setTheme($theme);
     }
 
-    /**
-     * @param int|null $id
-     * @return self
-     */
-    public function setId(?int $id): self
-    {
-        $this->id = $id;
-
-        return $this;
-    }
-
-    /**
-     * @return int|null
-     */
-    public function getId(): ?int
+    public function getId(): int
     {
         return $this->id;
-    }
-
-    /**
-     * @return string
-     */
-    public function getAlias(): string
-    {
-        return $this->alias;
-    }
-
-    /**
-     * @param string $alias
-     * @return self
-     * @throws \Assert\AssertionFailedException
-     */
-    public function setAlias(string $alias): self
-    {
-        Assertion::minLength($alias, self::MIN_ALIAS_LENGTH, 'User::alias');
-        Assertion::maxLength($alias, self::MAX_ALIAS_LENGTH, 'User::alias');
-        $this->alias = $alias;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    /**
-     * @param string $name
-     * @return self
-     * @throws \Assert\AssertionFailedException
-     */
-    public function setName(string $name): self
-    {
-        Assertion::minLength($name, self::MIN_ALIAS_LENGTH, 'User::name');
-        Assertion::maxLength($name, self::MAX_ALIAS_LENGTH, 'User::name');
-        $this->name = $name;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getEmail(): string
-    {
-        return $this->email;
-    }
-
-    /**
-     * @param string $email
-     * @return self
-     * @throws \Assert\AssertionFailedException
-     */
-    public function setEmail(string $email): self
-    {
-        // Email format validation cannot be done here until legacy form does not check it
-        Assertion::minLength($email, self::MIN_EMAIL_LENGTH, 'User::email');
-        Assertion::maxLength($email, self::MAX_EMAIL_LENGTH, 'User::email');
-        $this->email = $email;
-        return $this;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isAdmin(): bool
-    {
-        return $this->isAdmin;
-    }
-
-    /**
-     * @param bool $isAdmin
-     * @return self
-     */
-    public function setAdmin(bool $isAdmin): self
-    {
-        $this->isAdmin = $isAdmin;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getTheme(): string
-    {
-        return $this->theme;
-    }
-
-    /**
-     * @param string $theme
-     * @return self
-     * @throws \Assert\AssertionFailedException
-     */
-    public function setTheme(string $theme): self
-    {
-        Assertion::minLength($theme, self::MIN_THEME_LENGTH, 'User::theme');
-        Assertion::maxLength($theme, self::MAX_THEME_LENGTH, 'User::theme');
-        $this->theme = $theme;
-        return $this;
-    }
-
-    /**
-     * @return ContactTemplate|null
-     */
-    public function getContactTemplate(): ?ContactTemplate
-    {
-        return $this->contactTemplate;
-    }
-
-    /**
-     * @param ContactTemplate|null $contactTemplate
-     * @return self
-     */
-    public function setContactTemplate(?ContactTemplate $contactTemplate): self
-    {
-        $this->contactTemplate = $contactTemplate;
-
-        return $this;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isActivate(): bool
-    {
-        return $this->isActivate;
-    }
-
-    /**
-     * @param bool $isActivate
-     * @return self
-     */
-    public function setActivate(bool $isActivate): self
-    {
-        $this->isActivate = $isActivate;
-
-        return $this;
     }
 }
