+++ tests/php/Core/Application/Configuration/User/UseCase/PatchUser/PatchUserTest.php
@@ -81,10 +81,7 @@ it('tests the exception while searching for the user', function () {
 });
 
 it('tests the error message when there are no available themes', function () {
-    $user = new User('alias', 'name', 'email');
-    $user
-        ->setId(1)
-        ->setAdmin(true);
+    $user = new User(1, 'alias', 'name', 'email', true, 'light');
     $this->readUserRepository
         ->expects($this->once())
         ->method('findById')
@@ -106,10 +103,7 @@ it('tests the error message when there are no available themes', function () {
 });
 
 it('tests the error message when the given theme is not in the list of available themes', function () {
-    $user = new User('alias', 'name', 'email');
-    $user
-        ->setId(1)
-        ->setAdmin(true);
+    $user = new User(1, 'alias', 'name', 'email', true, 'light');
     $this->readUserRepository
         ->expects($this->once())
         ->method('findById')
@@ -130,10 +124,7 @@ it('tests the error message when the given theme is not in the list of available
 });
 
 it('tests the exception while searching for available themes', function () {
-    $user = new User('alias', 'name', 'email');
-    $user
-        ->setId(1)
-        ->setAdmin(true);
+    $user = new User(1, 'alias', 'name', 'email', true, 'light');
     $this->readUserRepository
         ->expects($this->once())
         ->method('findById')
@@ -156,10 +147,7 @@ it('tests the exception while searching for available themes', function () {
 });
 
 it('tests the exception while updating the theme of user', function () {
-    $user = new User('alias', 'name', 'email');
-    $user
-        ->setId(1)
-        ->setAdmin(true);
+    $user = new User(1, 'alias', 'name', 'email', true, 'light');
     $this->readUserRepository
         ->expects($this->once())
         ->method('findById')
